requirements.txt
Flask==3.0.3
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.2
marshmallow-sqlalchemy==1.1.0
Flask-JWT-Extended==4.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
Werkzeug==3.0.3
gunicorn==22.0.0
